feat(links): add form to add a new link
This commit adds a form to add a new link to the URL shortener. The form
is created using Django's form API and is displayed on the `Add a new
link` page. The form has fields for the link name, URL, and an optional
slug. The form is not yet functional, and the data is not saved to the
database.